Continued work on the routine.
It now starts from a specified starting location and then finds 9 data points at at length down the cone.
These data points are used to fit a circle.
The radius of the fit circle is used to predict where the application should move to down the cone.
The ultimate goal is to find the location down the cone that has a specified diameter.